update queing
git-svn-id: https://build.archivesportaleurope.net/svn/apex/trunk@175 9fd842fd-667d-44da-918d-ff91420b3ae1

+++ Dashboard/src/main/java/eu/apenet/dashboard/services/ead/DeleteTask.java
@@ -1,29 +1,48 @@
 package eu.apenet.dashboard.services.ead;
 
 import java.util.Properties;
-import java.util.Set;
 
+import eu.apenet.commons.exceptions.APEnetException;
 import eu.apenet.commons.utils.APEnetUtilities;
 import eu.apenet.dashboard.utils.ContentUtils;
 import eu.apenet.persistence.dao.EadDAO;
-import eu.apenet.persistence.dao.WarningsDAO;
 import eu.apenet.persistence.factory.DAOFactory;
 import eu.apenet.persistence.vo.Ead;
-import eu.apenet.persistence.vo.Warnings;
+import eu.apenet.persistence.vo.EuropeanaState;
+import eu.apenet.persistence.vo.FindingAid;
 
 public class DeleteTask extends AbstractEadTask {
 
 	@Override
+	protected String getActionName() {
+		return "delete";
+	}
+
+	@Override
 	protected void execute(Ead ead, Properties properties) throws Exception {
-		EadDAO eadDAO = DAOFactory.instance().getEadDAO();
-        WarningsDAO warningsDao = DAOFactory.instance().getWarningsDAO();
-        Set<Warnings> warningsList = ead.getWarningses();
-        for (Warnings warnings : warningsList) {
-            logger.debug("Removing warnings for EAD which EADID is " + ead.getEadid());
-            warningsDao.deleteSimple(warnings);
-        }
-        eadDAO.deleteSimple(ead);
-        ContentUtils.deleteFile(APEnetUtilities.getConfig().getRepoDirPath() + ead.getPathApenetead());
+		boolean valid = false;
+		if (!ead.isPublished()) {
+			if (ead instanceof FindingAid) {
+				FindingAid findingAid = (FindingAid) ead;
+				if (EuropeanaState.NOT_CONVERTED.equals(findingAid.getEuropeana())) {
+					valid =true;
+				}
+			}else{
+				valid =true;
+			}
+		}
+		if (valid) {
+			try {
+				ContentUtils.deleteFile(APEnetUtilities.getConfig().getRepoDirPath() + ead.getPathApenetead());
+				EadDAO eadDAO = DAOFactory.instance().getEadDAO();
+				eadDAO.delete(ead);
+				logAction(ead, true);
+			} catch (Exception e) {
+				logAction(ead, false);
+				throw new APEnetException("Could not delete the file with ID: " + ead.getId(), e);
+			}
+		}
+
 	}
 
 }
